Updated copyright info to 2012.
Changed the properties of HKItem to nonatomic to improve performance.

Added an HK_LAZY_INIT compiler option for HKFile and HKFolder. With it enabled, the properties like name, nameExtension, kind, fileType, size, and type aren't set until they are asked for. This should help performance dramatically.

Added an #HK_USE_BLOCKS compiler option to several classes to attempt to use NSEnumerationConcurrent enumeration to help improve performance. Had mixed results.



git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.5@90 c3d3028b-b5cd-4c35-8691-6fb8b2347b27

+++ HLKit/HKGCFFile.mm
@@ -3,7 +3,7 @@
 //  HLKit
 //
 //  Created by Mark Douma on 10/27/2010.
-//  Copyright (c) 2009-2011 Mark Douma LLC. All rights reserved.
+//  Copyright (c) 2009-2012 Mark Douma LLC. All rights reserved.
 //
 
 #import <HLKit/HKGCFFile.h>
@@ -43,7 +43,7 @@
 					
 					
 					if (static_cast<CGCFFile *>(_privateData)->GetAttribute(HL_GCF_PACKAGE_VERSION, versionAttr)) {
-						version = [[NSString stringWithFormat:@"1.%lu", versionAttr.Value.UnsignedInteger.uiValue] retain];
+						version = [[NSString stringWithFormat:@"1.%lu", (unsigned long)versionAttr.Value.UnsignedInteger.uiValue] retain];
 					}
 					if (static_cast<CGCFFile *>(_privateData)->GetAttribute(HL_GCF_PACKAGE_ID, packageIDAttr)) {
 						packageID = (NSUInteger)packageIDAttr.Value.UnsignedInteger.uiValue;
@@ -85,11 +85,11 @@
 - (NSString *)description {
 	NSMutableString *description = [NSMutableString stringWithString:@""];
 	[description appendFormat:@"\tfilePath == %@\n", filePath];
-	[description appendFormat:@"\tpackageID == %lu\n", packageID];
-	[description appendFormat:@"\tblockSize == %lu\n", blockSize];
-	[description appendFormat:@"\ttotalBlockCount == %lu\n", totalBlockCount];
-	[description appendFormat:@"\tusedBlockCount == %lu\n", usedBlockCount];
-	[description appendFormat:@"\tlastVersionPlayed == %lu\n", lastVersionPlayed];
+	[description appendFormat:@"\tpackageID == %lu\n", (unsigned long)packageID];
+	[description appendFormat:@"\tblockSize == %lu\n", (unsigned long)blockSize];
+	[description appendFormat:@"\ttotalBlockCount == %lu\n", (unsigned long)totalBlockCount];
+	[description appendFormat:@"\tusedBlockCount == %lu\n", (unsigned long)usedBlockCount];
+	[description appendFormat:@"\tlastVersionPlayed == %lu\n", (unsigned long)lastVersionPlayed];
 	return [NSString stringWithFormat:@"%@", description];
 }
